pyproject.toml
[project]
name = "gen-subtitle"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "argostranslate>=1.11.0",
    "deepl>=1.28.0",
    "python-dotenv>=1.2.2",
]
